Remove Microsoft.Framework.Project
