+++ XRE.sln
@@ -36,8 +36,6 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Command
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.DesignTimeHost", "src\Microsoft.Framework.DesignTimeHost\Microsoft.Framework.DesignTimeHost.kproj", "{B8208D93-75C4-44BF-80BC-DEB9678FA05E}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Project", "src\Microsoft.Framework.Project\Microsoft.Framework.Project.kproj", "{AE99F965-68E2-4E7A-947A-0D2C61992792}"
-EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "HelloWorld", "samples\HelloWorld\HelloWorld.kproj", "{BF98054D-21EB-46B2-994B-007B145D32A1}"
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "HelloShared", "samples\HelloShared\HelloShared.kproj", "{87F66056-D3D9-47E0-8F9B-DF385CCF03AD}"
@@ -219,20 +217,6 @@ Global
 		{B8208D93-75C4-44BF-80BC-DEB9678FA05E}.Release|Win32.ActiveCfg = Release|Any CPU
 		{B8208D93-75C4-44BF-80BC-DEB9678FA05E}.Release|x64.ActiveCfg = Release|Any CPU
 		{B8208D93-75C4-44BF-80BC-DEB9678FA05E}.Release|x86.ActiveCfg = Release|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Release|Win32.ActiveCfg = Release|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Release|x64.ActiveCfg = Release|Any CPU
-		{AE99F965-68E2-4E7A-947A-0D2C61992792}.Release|x86.ActiveCfg = Release|Any CPU
 		{BF98054D-21EB-46B2-994B-007B145D32A1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BF98054D-21EB-46B2-994B-007B145D32A1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BF98054D-21EB-46B2-994B-007B145D32A1}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -560,7 +544,6 @@ Global
 		{7829F696-AFC4-4011-B9DE-6F1C24846D67} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
 		{C46A8C00-CD50-4478-8639-6A6CF8CDD05B} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
 		{B8208D93-75C4-44BF-80BC-DEB9678FA05E} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
-		{AE99F965-68E2-4E7A-947A-0D2C61992792} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
 		{BF98054D-21EB-46B2-994B-007B145D32A1} = {3F68EE40-3C6A-4FD1-B057-0427A9BF88BE}
 		{87F66056-D3D9-47E0-8F9B-DF385CCF03AD} = {3F68EE40-3C6A-4FD1-B057-0427A9BF88BE}
 		{0D752D1C-3985-4E27-92C4-0BEE72B1F58D} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
