+++ scripts/kpm.cmd
@@ -2,7 +2,7 @@
 SETLOCAL
 SET ERRORLEVEL=
 
-"%~dp0dotnet" %DOTNET_OPTIONS% --lib "%~dp0lib\Microsoft.Framework.Project" "%~dp0lib\Microsoft.Framework.PackageManager\Microsoft.Framework.PackageManager.dll" %*
+"%~dp0dotnet" %DOTNET_OPTIONS% "%~dp0lib\Microsoft.Framework.PackageManager\Microsoft.Framework.PackageManager.dll" %*
 
 exit /b %ERRORLEVEL%
 ENDLOCAL
