+++ src/DependencyAnalyzer/Util/DependencyFinder.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Microsoft.Framework.Project;
 using Microsoft.Framework.Runtime;
 using NuGet;
