+++ src/DependencyAnalyzer/project.json
@@ -1,7 +1,6 @@
 {
     "dependencies": {
         "Microsoft.Framework.Runtime": "1.0.0-*",
-        "Microsoft.Framework.Project": "1.0.0-*",
         "Microsoft.Framework.CommandLineUtils": { "version": "1.0.0-*", "type": "build" }
     },
     "commands": {
