+++ src/Microsoft.Framework.PackageManager/Crossgen/CrossgenOptions.cs
@@ -1,13 +1,9 @@
-// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
+﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Microsoft.Framework.Project
+namespace Microsoft.Framework.PackageManager.Crossgen
 {
     public class CrossgenOptions
     {
@@ -16,7 +12,7 @@ namespace Microsoft.Framework.Project
         public string RuntimePath { get; set; }
 
         public IEnumerable<string> InputPaths { get; set; }
-        
+
         public bool Symbols { get; set; }
 
         public bool Partial { get; set; }
