+++ src/Microsoft.Framework.PackageManager/Program.cs
@@ -6,9 +6,10 @@ using System.IO;
 using System.Net;
 using System.Reflection;
 using System.Threading;
-using Microsoft.Framework.PackageManager.Packages;
-using Microsoft.Framework.PackageManager.List;
 using Microsoft.Framework.PackageManager.Bundle;
+using Microsoft.Framework.PackageManager.Crossgen;
+using Microsoft.Framework.PackageManager.List;
+using Microsoft.Framework.PackageManager.Packages;
 using Microsoft.Framework.Runtime;
 using Microsoft.Framework.Runtime.Common.CommandLine;
 
@@ -441,6 +442,36 @@ namespace Microsoft.Framework.PackageManager
 
                     var command = new DependencyListCommand(options);
                     return command.Execute();
+                });
+            });
+
+            app.Command("crossgen", c =>
+            {
+                c.Description = "Do CrossGen";
+
+                var optionIn = c.Option("--in <INPUT_DIR>", "Input directory", CommandOptionType.MultipleValue);
+                var optionOut = c.Option("--out <OUTOUT_DIR>", "Output directory", CommandOptionType.SingleValue);
+                var optionExePath = c.Option("--exePath", "Exe path", CommandOptionType.SingleValue);
+                var optionRuntimePath = c.Option("--runtimePath <PATH>", "Runtime path", CommandOptionType.SingleValue);
+                var optionSymbols = c.Option("--symbols", "Use symbols", CommandOptionType.NoValue);
+                var optionPartial = c.Option("--partial", "Allow partial NGEN", CommandOptionType.NoValue);
+                c.HelpOption("-?|-h|--help");
+
+                c.OnExecute(() =>
+                {
+                    var crossgenOptions = new CrossgenOptions();
+                    crossgenOptions.InputPaths = optionIn.Values;
+                    crossgenOptions.RuntimePath = optionRuntimePath.Value();
+                    crossgenOptions.CrossgenPath = optionExePath.Value();
+                    crossgenOptions.Symbols = optionSymbols.HasValue();
+                    crossgenOptions.Partial = optionPartial.HasValue();
+
+                    var gen = new CrossgenManager(crossgenOptions);
+                    if (!gen.GenerateNativeImages())
+                    {
+                        return -1;
+                    }
+                    return 0;
                 });
             });
