--- src/Microsoft.Framework.Project/project.json
@@ -1,40 +1,0 @@
-{
-    "version": "1.0.0-*",
-    "compilationOptions" : { "define" : ["TRACE"], "warningsAsErrors": true },
-    "dependencies": {
-        "Microsoft.Framework.CommandLineUtils" : { "version": "1.0.0-*", "type": "build" },
-        "Microsoft.Framework.Runtime" : "1.0.0-*",
-        "Microsoft.Framework.Runtime.Common" : { "version": "1.0.0-*", "type": "build" },
-        "Microsoft.Framework.Runtime.Interfaces" : { "version": "1.0.0-*", "type": "build" }
-    },
-    "frameworks" : {
-        "aspnet50" : {
-            "frameworkAssemblies": {
-                "System.Collections" : "",
-                "System.IO" :  "",
-                "System.IO.Compression" : "",
-                "System.Runtime" : "",
-                "System.Xml": "",
-                "System.Xml.Linq": ""
-            }
-        },
-        "aspnetcore50" : { 
-            "dependencies": {
-                "System.Collections": "4.0.10-beta-*",
-                "System.ComponentModel": "4.0.0-beta-*",
-                "System.Console": "4.0.0-beta-*",
-                "System.Diagnostics.Debug": "4.0.10-beta-*",
-                "System.Diagnostics.Process": "4.0.0-beta-*",
-                "System.IO": "4.0.10-beta-*",
-                "System.IO.Compression": "4.0.0-beta-*",
-                "System.IO.FileSystem": "4.0.0-beta-*",
-                "System.Linq": "4.0.0-beta-*",
-                "System.Reflection": "4.0.10-beta-*",
-                "System.Runtime": "4.0.20-beta-*",
-                "System.Runtime.Extensions": "4.0.10-beta-*",
-                "System.Runtime.InteropServices": "4.0.20-beta-*",
-                "System.Threading.Tasks": "4.0.10-beta-*"
-            }
-        }
-    }
-}
